Initial redux refactor
