+++ src/actions/index.js
@@ -1,12 +1,22 @@
-import * as Rx from '@reactivex/rxjs';
-
 import { getJSON } from '../api';
 
-export const actions = new Rx.Subject();
+export function readJSON(filePath) {
+  return (subject) => {
+    getJSON(filePath)
+      .then((data) =>
+        subject.next({
+          type: 'READ_JSON',
+          data,
+        })
+      );
+  };
+}
 
-export const readJSON = (filePath) =>
-  Rx.Observable.fromPromise(getJSON(filePath))
-    .subscribe(data => actions.next(Object.assign({}, { type: 'READ_JSON' }, { data })));
-
-export const updateCell = (notebook, index, cell) =>
-  actions.next(Object.assign({}, { type: 'UPDATE_CELL' }, { notebook, index, cell }));
+export function updateCell(notebook, index, cell) {
+  return {
+    type: 'UPDATE_CELL',
+    notebook,
+    index,
+    cell,
+  };
+}
